feat: try to upload/download docker image tar ball and load into docker between steps (#755)
* feat: try to upload/download docker image tar ball and load into docker


* feat: try output type tar


* fix: mkdir


* fix: save tar


* feat: download pre-built image tarballs


* fix: integration test


* fix: integration test


* fix: integration test


* refactor: rename job


* refactor: rename job


* fix: try to fix Permission denied issue


* fix: try to fix Permission denied issue


* refactor: rename job step


* fix: IMAGE_PULL_POLICY


* refactor: add env variable to control if create local registry


* refactor: k3d import all images in a batch


* refactor: add log for debugging


* refactor: add log for debugging


* test: k3d try to load image from docker instead of tarball


* test: k3d try to load image from docker instead of tarball


* test: use local registry


* ci: revert a change


* ci: mysql image


* fix: scripts


* revert: k3d-with-registry.sh


* revert: k3d-with-registry.sh


* ci: ubuntu 22.04 for integration test


* fix: mysql startup parameters


* fix: mysql startup parameters


* fix: mysql service name


* refactor: add debug log


* refactor: remove ns from svc name


* fix: K3D_FIX_DNS=0


* revert: mysql service name


* fix: use kind for integration test


* revert: deploy-mysql.sh


* build: DOCKER_GO_VERSION


* revert: gorm version


* refactor: workflow


* test: env


* test: remove CTR_REGISTRY from env for integration test


* test: remove IMAGE_PULL_POLICY from env for integration test


* test: use github.com/go-sql-driver/mysql as mysql driver


* test: use github.com/go-sql-driver/mysql as mysql driver


* test: try to resolve IP of mysql first


* test: increase CI_WAIT_FOR_OK_SECONDS


* test: add logs for debugging


* test: add logs for debugging


* test: add logs for debugging


* test: increase CI_MAX_ITERATIONS_THRESHOLD


* test: add log for debugging


* test: adjust config and log


---------

+++ scripts/k3d-with-registry.sh
@@ -12,7 +12,9 @@ K3D_GATEWAY_API_ENABLE="${K3D_GATEWAY_API_ENABLE:-false}"
 K3D_FLB_ENABLE="${K3D_FLB_ENABLE:-false}"
 K3D_SERVICELB_ENABLE="${K3D_SERVICELB_ENABLE:-false}"
 K3D_IMAGE="${K3D_IMAGE:-rancher/k3s:v1.25.16-k3s4}"
-CI_INTEGRATION_TEST="${CI_INTEGRATION_TEST:-false}"
+CREATE_LOCAL_REGISTRY="${CREATE_LOCAL_REGISTRY:-true}"
+NUMBER_OF_K3D_SERVERS="${NUMBER_OF_K3D_SERVERS:-1}"
+NUMBER_OF_K3D_AGENTS="${NUMBER_OF_K3D_AGENTS:-0}"
 
 # shellcheck disable=SC2086
 jq_cluster_exists=".[] | select(.name == \"${K3D_CLUSTER_NAME}\")"
@@ -22,38 +24,43 @@ if [ "${num_of_cluster_exists}" = "1" ] ; then
   exit 0
 fi
 
-k3d_network='fsm'
+#k3d_network='fsm'
 k3d_prefix='k3d'
 reg_name='registry.localhost'
 final_reg_name="${k3d_prefix}-${reg_name}"
 reg_port='5000'
 
-# create registry container unless it already num_of_exists
-jq_reg_exists=".[] | select(.name == \"${final_reg_name}\")"
-jq_reg_running=".[] | select(.name == \"${final_reg_name}\" and .State.Running == true)"
-num_of_exists=$(k3d registry list -o json | jq "${jq_reg_exists}" | jq -s 'length')
-if [ "${num_of_exists}" = '0' ]; then
-  # no k3d managed registry found, kill any running registry container and create a new one
-  #shellcheck disable=SC2046
-  container_id=$(docker ps --format json | jq -r 'select(.Image == "registry:2" and .State == "running") | .ID')
-  if [ "${container_id}" != "" ]; then
-    docker kill "${container_id}"
-  fi
-  k3d registry create "${reg_name}" --port "127.0.0.1:${reg_port}"
-else
-  num_of_running=$(k3d registry list -o json | jq "${jq_reg_running}" | jq -s 'length')
-  if [ "${num_of_running}" = '0' ]; then
+if [ "${CREATE_LOCAL_REGISTRY}" = "true" ]; then
+  # create registry container unless it already num_of_exists
+  jq_reg_exists=".[] | select(.name == \"${final_reg_name}\")"
+  jq_reg_running=".[] | select(.name == \"${final_reg_name}\" and .State.Running == true)"
+  num_of_exists=$(k3d registry list -o json | jq "${jq_reg_exists}" | jq -s 'length')
+  if [ "${num_of_exists}" = '0' ]; then
     # no k3d managed registry found, kill any running registry container and create a new one
-    k3d registry delete --all
     #shellcheck disable=SC2046
     container_id=$(docker ps --format json | jq -r 'select(.Image == "registry:2" and .State == "running") | .ID')
     if [ "${container_id}" != "" ]; then
       docker kill "${container_id}"
     fi
     k3d registry create "${reg_name}" --port "127.0.0.1:${reg_port}"
+  else
+    num_of_running=$(k3d registry list -o json | jq "${jq_reg_running}" | jq -s 'length')
+    if [ "${num_of_running}" = '0' ]; then
+      # no k3d managed registry found, kill any running registry container and create a new one
+      k3d registry delete --all
+      #shellcheck disable=SC2046
+      container_id=$(docker ps --format json | jq -r 'select(.Image == "registry:2" and .State == "running") | .ID')
+      if [ "${container_id}" != "" ]; then
+        docker kill "${container_id}"
+      fi
+      k3d registry create "${reg_name}" --port "127.0.0.1:${reg_port}"
+    fi
   fi
 fi
 
+reg_config=""
+
+if [ "${CREATE_LOCAL_REGISTRY}" = "true" ]; then
 reg_config="
 registries:
   use:
@@ -64,26 +71,6 @@ registries:
         endpoint:
           - http://${final_reg_name}:${reg_port}
 "
-
-if [ "${CI_INTEGRATION_TEST}" = "true" ]; then
-reg_config="
-registries:
-  config: |
-    mirrors:
-      ghcr.io:
-      docker.io:
-    configs:
-      ghcr.io:
-        auth:
-          username: ${GITHUB_REPO_OWNER}
-          password: ${GITHUB_TOKEN}
-      docker.io:
-        tls:
-          insecure_skip_verify: true
-        auth:
-          username: ${DOCKER_USER}
-          password: ${DOCKER_PASS}
-"
 fi
 
 # create cluster
@@ -92,10 +79,9 @@ apiVersion: k3d.io/v1alpha5
 kind: Simple
 metadata:
   name: ${K3D_CLUSTER_NAME}
-servers: 1
-agents: 0
+servers: ${NUMBER_OF_K3D_SERVERS}
+agents: ${NUMBER_OF_K3D_AGENTS}
 image: ${K3D_IMAGE}
-network: ${k3d_network}
 ${reg_config}
 ports:
   - port: 80:80
